Update to 1.16

+++ src/main/java/net/urbanmc/mcinfected/object/GamePlayer.java
@@ -108,10 +108,10 @@ public class GamePlayer {
 	}
 
 	public String getKDR() {
-
 		double kdr;
+
 		if (deaths != 0)
-			kdr = kills / deaths;
+			kdr = (double) kills / deaths;
 		else
 			kdr = kills;
